Melomind battery percentage moved to MelomindDevice class
